Message position is decided
